Reduce the chance of random failure in invalid_block_sync_test.
The cause is that when we resend a GetBlocks request, the RequestBatcher
does not take previous information of failing peers into consideration.

Only one of the two peers can provide all blocks, but it's possible that
the node always choose the wrong one to request. Increasing the timeout
to 120 s is a quick fix to reduce the failure probability to about
1/2^15 (the delay of a request is increased by 1 second for each failure),
which is small enough for now.